+++ src/tile.c
@@ -264,7 +264,7 @@ void Tile_RemoveFogInRadius(tile32 tile, uint16 radius)
 
 			if (Tile_GetDistanceRoundedUp(tile, t) > radius) continue;
 
-			Map_UnveilTile(packed, g_global->playerHouseID);
+			Map_UnveilTile(packed, (uint8)g_global->playerHouseID);
 		}
 	}
 }
